BachelorThesis/webapp/requirements.txt
flask==2.1.1
flask_cors==3.0.10
pymongo==4.1.0
dnspython==2.2.1